+++ src/loop/expr_intf.ml
@@ -3,7 +3,7 @@
 (** Expressions
 
     This modules defines the smallest signatures for expressions that allow
-    to isntantiates the {Pipes.Make} functor. *)
+    to instantiates the {Pipes.Make} functor. *)
 
 module type S = sig
 
@@ -36,3 +36,63 @@ module type Print = sig
   val formula : Format.formatter -> formula -> unit
 
 end
+
+module type Export = sig
+
+  include S
+
+  type 'a tag
+
+  module Tags : sig
+
+    val named : string tag
+
+  end
+
+  module Ty : sig
+    type t = ty
+    module Var : sig
+      include Dolmen_intf.Id.Scope
+      with type t = ty_var
+       and type path := Dolmen_std.Path.t
+    end
+    module Const : sig
+      include Dolmen_intf.Id.Scope
+      with type t = ty_cst
+       and type path := Dolmen_std.Path.t
+    end
+  end
+
+  module Term : sig
+    type t = term
+
+    val neg : t -> t
+    (** Logical negation. *)
+
+    val _or : t list -> t
+    (** Disjunction. *)
+
+    val of_cst : term_cst -> t
+    (** Create a formula out of a constant. *)
+
+    module Var : sig
+      include Dolmen_intf.Id.Scope
+      with type t = term_var
+       and type path := Dolmen_std.Path.t
+    end
+
+    module Const : sig
+      include Dolmen_intf.Id.Scope
+      with type t = term_cst
+       and type path := Dolmen_std.Path.t
+
+      val mk : Dolmen_std.Path.t -> ty -> t
+      (** Create a constant symbol *)
+
+      val get_tag : t -> 'a tag -> 'a option
+      (** Return the value bound to a tag (if any). *)
+    end
+  end
+
+end
+