+++ src/typecheck/intf.ml
@@ -674,6 +674,7 @@ module type Formulas = sig
   type 'a ret = {
     implicit_decls : decl list;
     implicit_defs : def list;
+    recursive : bool;
     result : 'a;
   }